% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fex}
\alias{fex}
\title{File extension}
\usage{
fex(x)
}
\arguments{
\item{x}{character vector giving file paths.}
}
\value{
A character vector of the same length as \code{x} containing the file
extensions without the leading dot. Files without an extension return an
empty string ("").
}
\description{
file extension
}
\note{
source tools::file_ext
}
\examples{
\dontrun{
# Do not run
}
}
\seealso{
\code{tools::file_ext()}

Other helpers: 
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
