% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.invfile}
\alias{print.invfile}
\alias{invfile}
\alias{summary.invfile}
\alias{plot.invfile}
\title{Methods for objects with class `invfile`}
\usage{
\method{print}{invfile}(x, ...)

\method{summary}{invfile}(object, ...)

\method{plot}{invfile}(x, time = "timeUTC", ...)
}
\arguments{
\item{x}{Object with class "invfile"}

\item{...}{plot arguments}

\item{object}{invfile}

\item{time}{x axis column (time)}
}
\value{
Objects of class "invfile"
}
\description{
Methods for objects with class `invfile`
}
\examples{
\dontrun{
#do not run
}
}
