% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_addtime_obs.R
\name{obs_addltime}
\alias{obs_addltime}
\title{add local time}
\usage{
obs_addltime(
  dt,
  timeUTC = "timeUTC",
  utc2lt = "site_utc2lst",
  longitude = "longitude",
  tz = "UTC",
  timeonly = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{timeUTC}{Character indicating the Time column as POSIXct}

\item{utc2lt}{Character indicating the integer column to convert to local time
if available}

\item{longitude}{Character indicating the column with the lingitude}

\item{tz}{Timezone, default "UTC"}

\item{timeonly}{return only local_time column}
}
\value{
data.table with local time columns
}
\description{
Calculate an approximation of local hour
}
\note{
time depending n longitude is by John Miller (GML/NOAA)
}
\examples{
\dontrun{
# Do not run
}
}
\seealso{
Other time: 
\code{\link{obs_addmtime}()},
\code{\link{obs_addstime}()},
\code{\link{obs_addtime}()}
}
\concept{time}
