% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_addtime_obs.R
\name{obs_addmtime}
\alias{obs_addmtime}
\title{Add matlab time}
\usage{
obs_addmtime(time, origin = "1970-01-01", tz = "UTC")
}
\arguments{
\item{time}{numeric time from MATLAB}

\item{origin}{default "1970-01-01"}

\item{tz}{Timezone, default "UTC"}
}
\value{
return the same data.frame adding solar time
}
\description{
This function add timeUTC as POSIX class,
local time and ending sampling time
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <-  obs_summary(obs)
dt <- obs_read(index)
dt <- obs_addtime(dt)
}
}
\seealso{
Other time: 
\code{\link{obs_addltime}()},
\code{\link{obs_addstime}()},
\code{\link{obs_addtime}()}
}
\concept{time}
