% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_addtime_obs.R
\name{obs_addstime}
\alias{obs_addstime}
\title{Add solar time into obspack}
\usage{
obs_addstime(dt, verbose = TRUE)
}
\arguments{
\item{dt}{obspack data.table}

\item{verbose}{obspack data.table}
}
\value{
return the same data.frame adding solar time
}
\description{
This function add timeUTC as POSIX class,
local time and ending sampling time
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <-  obs_summary(obs)
dt <- obs_read(index)
dt <- obs_addtime(dt)
}
}
\seealso{
Other time: 
\code{\link{obs_addltime}()},
\code{\link{obs_addtime}()}
}
\concept{time}
