% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate_obs.R
\name{obs_agg}
\alias{obs_agg}
\title{Aggregates ObsPack by time}
\usage{
obs_agg(
  dt,
  cols = c("year", "month", "day", "hour", "minute", "second", "time", "time_decimal",
    "value", "latitude", "longitude", "altitude", "pressure", "u", "v", "temperature",
    "type_altitude"),
  by = c("key_time", "site_code", "altitude_final", "type_altitude", "lab_1_abbr",
    "dataset_calibration_scale"),
  fn = "mean",
  verbose = TRUE
)
}
\arguments{
\item{dt}{obspack data.table}

\item{cols}{Character which defines columns to be aggregated}

\item{by}{String with columns to be aggregated}

\item{fn}{Function to be applied to the columns, default mean}

\item{verbose}{logical to show more information}
}
\value{
A data.frame with with an index obspack.
}
\description{
This function add timeUTC as POSIX class,
local time and ending sampling time
}
\note{
By default add column timeUTC based on the input column `key_time`
}
\examples{
\dontrun{
# Do not run
# dt <- obs_addtime(dt)
}
}
