% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_decimal_to_POSIX}
\alias{obs_decimal_to_POSIX}
\title{obs_decimal_to_POSIX}
\usage{
obs_decimal_to_POSIX(decimal.dates, tz = "UTC", legacy = TRUE)
}
\arguments{
\item{decimal.dates}{vecot with format YYYY.yyyy}

\item{tz}{time zone, "UTC" default}

\item{legacy}{logical, to use legacy code}
}
\value{
return list of time vars
}
\description{
return list of time vars
}
\examples{
{
\dontrun{
# Do not run
}}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()},
\code{\link{obs_interpret_udunits_time}()},
\code{\link{obs_julian}()},
\code{\link{obs_load_ncdf}()},
\code{\link{obs_make_hysplit_nc4}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
