% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_epoch_seconds_to_POSIX}
\alias{obs_epoch_seconds_to_POSIX}
\title{obs_epoch_seconds_to_POSIX}
\usage{
obs_epoch_seconds_to_POSIX(sec, tz = "UTC")
}
\arguments{
\item{sec}{seconds}

\item{tz}{time zone}
}
\value{
return dates
}
\description{
return seconds
}
\examples{
{
\dontrun{
# Do not run
obs_epoch_seconds_to_POSIX(1)
}}
}
\seealso{
Other helpers legacy time: 
\code{\link{obs_POSIX_to_decimal}()},
\code{\link{obs_POSIX_to_epoch_seconds}()},
\code{\link{obs_POSIX_to_idate}()},
\code{\link{obs_POSIX_to_julian}()},
\code{\link{obs_days_in_month}()},
\code{\link{obs_days_in_year}()},
\code{\link{obs_extract_dates}()},
\code{\link{obs_format_et}()},
\code{\link{obs_idate_to_POSIX}()},
\code{\link{obs_is_leap}()},
\code{\link{obs_itau_to_POSIX}()},
\code{\link{obs_julian_to_POSIX}()},
\code{\link{obs_length_POSIXlt}()},
\code{\link{obs_seconds_in_month}()},
\code{\link{obs_seconds_in_year}()},
\code{\link{obs_seq_midmon}()}
}
\concept{helpers legacy time}
