% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_find_missing_receptors.R
\name{obs_find_receptors}
\alias{obs_find_receptors}
\title{Compares expected receptors}
\usage{
obs_find_receptors(
  path,
  year,
  month,
  day,
  hour,
  minute,
  lat,
  lon,
  alt,
  out = paste0(tempfile(), ".csvy"),
  verbose = FALSE
)
}
\arguments{
\item{path}{String with path hwere are stored the footprints}

\item{year}{numeric number}

\item{month}{numeric number}

\item{day}{numeric number}

\item{hour}{numeric number}

\item{minute}{numeric number}

\item{lat}{numeric number}

\item{lon}{numeric number}

\item{alt}{numeric number}

\item{out}{outfile path.}

\item{verbose}{logical to show more information}
}
\value{
A data.frame with with expected footprints
}
\description{
This function creates a data.frame with the expected
footprints in .nc and compare with the actual footprints.
The idea is verify is there are missing footprints.
}
\examples{
\dontrun{
# do not run
p <- "/path/to/continuous/"
# here we have year/month/hysplit*.nc
x <- dt
dt <- obs_find_receptors(p, year, month....)
}
}
