% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_obs_convolve.R
\name{obs_foot}
\alias{obs_foot}
\title{obs_footprints}
\usage{
obs_foot(
  foot_path = "AAA",
  name_foot = "foot1",
  flon = "foot1lon",
  flat = "foot1lat",
  time_foot,
  fn = NULL,
  as_list = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{foot_path}{path for footprint (length 1).}

\item{name_foot}{name of the footprint variable in the NetCDF file.}

\item{flon}{name of lons}

\item{flat}{name of lats}

\item{time_foot}{Time of the footprints (in the name file) or third dimension}

\item{fn}{string with function to aggregate convolved fluxes, e.g. `mean`, `sum`, `max`, etc.}

\item{as_list}{Logical, to return list of arrays}

\item{verbose}{Logical, to display more information}
}
\value{
An array of footprints, or a list of footprints, latitudes, longitudes and time.
}
\description{
This function returns a arrays (or list of arrays) of convolved footpritns with flux
}
\examples{
\dontrun{
# do not run
}
}
