% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_footname}
\alias{obs_footname}
\title{Expected footprint name}
\usage{
obs_footname(
  time = NULL,
  year,
  month,
  day,
  hour,
  minute,
  second,
  lat,
  lon,
  alt,
  fullpath = FALSE,
  out,
  ...
)
}
\arguments{
\item{time}{POSIXct time to extract time variblaes}

\item{year}{numeric number}

\item{month}{numeric number}

\item{day}{numeric number}

\item{hour}{numeric number}

\item{minute}{numeric number}

\item{second}{numeric number, optional}

\item{lat}{numeric number}

\item{lon}{numeric number}

\item{alt}{numeric number}

\item{fullpath}{Logical, to add or not YYYY/MO/hysplit to id}

\item{out}{outfile path.}

\item{...}{data.table::fwrite arguments.}
}
\description{
return the expected name for the NetCDF footprint
}
\note{
source https://stackoverflow.com/a/47015304/2418532

# IMPORTANT!!!
# This function will generate the expected NetCDF file name.
# It assumes that the name was generated under the following considerations:
# time variables (year, month, day, etc) have a format of two digits, eg "0.1"
# latitude and longitude have been round with 4 decimals
# The format for latitude is 2 integers, a point and 4 decimals
# The format for longitude is 3 integers, a point and 4 decimals
}
\examples{
\dontrun{
# Do not run
obs_footname(time = Sys.time(),
             second = data.table::second(Sys.Time()),
             lat = 0,
             lon = 0,
             alt = 0)
obs_footname(year = 2020,
             month = 12,
             day = 30,
             hour = 9,
             minute = 54,
             lat = 3.2133,
             lon = 30.9131,
             alt = 497,
             fullpath = TRUE)
obs_footname(year = 2020,
             month = 12,
             day = 30,
             hour = 9,
             minute = 54,
             lat = 1,
             lon = -130.9131,
             alt = 497,
             fullpath = TRUE)
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
