% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robspack-deprecated.R
\name{obs_format}
\alias{obs_format}
\title{Formatting data}
\usage{
obs_format()
}
\arguments{
\item{dt}{`data.table`}

\item{spf}{columns to be formatted with `sprintf`}

\item{spffmt}{format to be applied to spf}

\item{rnd}{columns to be round}

\item{rndn}{Round number to be applied to `rnd`}

\item{spfrnd}{Logical, sprintf `rnd` columns after being round ?}

\item{out}{outfile path used by data.table::fwrite.}

\item{...}{data.table::fwrite arguments.}
}
\description{
return data.frame with formatted fields
}
\note{
source https://stackoverflow.com/a/47015304/2418532
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
