% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_freq}
\alias{obs_freq}
\title{return numeric vector in intervals}
\usage{
obs_freq(x, freq, ...)
}
\arguments{
\item{x}{numeric, longer than `freq`}

\item{freq}{numeric vector}

\item{...}{findInterval arguments}
}
\value{
numeric vector
}
\description{
return numeric vector in intervals
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
