% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_obs_hysplit.R
\name{obs_grid}
\alias{obs_grid}
\title{obs_grid
Create grid for footprints}
\usage{
obs_grid(
  numpix.x = 70,
  numpix.y = 70,
  lon.ll = 10,
  lat.ll = -30,
  lon.res = 1,
  lat.res = 1
)
}
\arguments{
\item{numpix.x}{number of pixels in x direction}

\item{numpix.y}{number of pixels in y direction}

\item{lon.ll}{lower left corner of grid}

\item{lat.ll}{lower left corner of grid}

\item{lon.res}{resolution in degrees longitude}

\item{lat.res}{resolution in degrees latitude}
}
\value{
A grid
}
\description{
obs_grid
Create grid for footprints
}
\note{
name it is not obs_hysplit_grid because we can create define grid in CONTROL
}
\examples{
{
# Do not run
obs_hysplit_grid(control_file)

}
}
