% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_obs_hysplit.R
\name{obs_hysplit_ascdata}
\alias{obs_hysplit_ascdata}
\title{obs_hysplit_ascdata}
\usage{
obs_hysplit_ascdata(
  llc = c(-90, -180),
  spacing = c(1, 1),
  n = c(180, 360),
  landusecat = 2,
  rough = 0.2,
  bdyfiles = "../bdyfiles/",
  ascdata = "ASCDATA.CFG"
)
}
\arguments{
\item{llc}{Lower left corner, default c(-90.0, -180.0)}

\item{spacing}{spacing in degress, default c(1.0, 1.0)}

\item{n}{number of data points, default c(180, 360)}

\item{landusecat}{land use category, default 2}

\item{rough}{default roughness length (meters), default 0.2}

\item{bdyfiles}{directory location of the data files,
default '../bdyfiles/'}

\item{ascdata}{file, default ASCDATA.CFG}
}
\value{
A ASCDATA.CFG file
}
\description{
This function creates a ASCDATA.CFG file for HYSPLIT model.
}
\examples{
{
# Do not run
ascdata_file <- tempfile()
obs_hysplit_ascdata(ascdata = ascdata_file)
cat(readLines(ascdata_file), sep =  "\n")
}
}
