% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_obs_cdump.R
\name{obs_hysplit_cdump2nc}
\alias{obs_hysplit_cdump2nc}
\title{Convert a HYSPLIT cdump file to CF-compliant NetCDF}
\usage{
obs_hysplit_cdump2nc(
  cdump,
  nc_out = NULL,
  drange = NULL,
  century = NULL,
  massunit = "1",
  pollutant = "pollutant",
  height_reference = c("sea_level", "ground_level"),
  control = NULL,
  setup = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdump}{Path to the binary cdump file.}

\item{nc_out}{Path of the NetCDF file to create. Default
\code{NULL} writes \code{<cdump>.nc} next to the source file.}

\item{drange}{Optional period to load, see
\code{\link{obs_hysplit_cdump_read}}.}

\item{century}{Integer century for the 2-digit years, see
\code{\link{obs_hysplit_cdump_read}}.}

\item{massunit}{Mass unit of the concentrations. Default
\code{"1"}.}

\item{pollutant}{Species name used to build the CF
\code{standard_name}. Default \code{"pollutant"}.}

\item{height_reference}{Reference of the level heights, see
\code{\link{obs_hysplit_cdump_nc}}.}

\item{control}{Path of the CONTROL file to embed, see
\code{\link{obs_hysplit_cdump_nc}}.}

\item{setup}{Path of the SETUP.CFG file to embed, see
\code{\link{obs_hysplit_cdump_nc}}.}

\item{verbose}{Logical, to display more information.}
}
\value{
The path of the created NetCDF file, invisibly.
}
\description{
One-call wrapper around \code{\link{obs_hysplit_cdump_read}} and
\code{\link{obs_hysplit_cdump_nc}}: reads the binary HYSPLIT
concentration file and writes the equivalent CF-compliant
NetCDF with dimensions \code{time}, \code{z}, \code{y}, \code{x}.
}
\examples{
\dontrun{
obs_hysplit_cdump2nc("cdump")
}
}
\seealso{
\code{\link{obs_hysplit_cdump_read}},
  \code{\link{obs_hysplit_cdump_nc}}
}
