% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_obs_hysplit.R
\name{obs_hysplit_control}
\alias{obs_hysplit_control}
\title{obs_hysplit_control}
\usage{
obs_hysplit_control(
  df,
  year,
  month,
  day,
  hour,
  minute,
  lat,
  lon,
  alt,
  start_time = NULL,
  nlocations = 1,
  duration = -240,
  vertical_motion = 5,
  top_model_domain = 10000,
  met = c("hrrr", "nams", "gfs0p25"),
  nmet = abs(duration/24) + 1,
  metpath = c("/work/noaa/lpdm/metfiles/hrrr/", "/work/noaa/lpdm/metfiles/nams/",
    "/work/noaa/lpdm/metfiles/gfs0p25/"),
  metformat = c(hrrr = "\%Y\%m\%d_hrrr", nams = "\%Y\%m\%d_hysplit.t00z.namsa", gfs0p25 =
    "\%Y\%m\%d_gfs0p25", era5 = "ERA5_\%Y\%m\%d.ARL"),
  ngases = 1,
  gas = "Foot",
  emissions_rate = 0,
  hour_emissions = 0.01,
  release_start = NULL,
  nsim_grids = 1,
  center_conc_grids = c(0, 0),
  grid_spacing = c(0.1, 0.1),
  grid_span = c(30, 30),
  nconc = "cdump",
  nvert_levels = 1,
  height_vert_levels = 50,
  sampling_start_time = c(0, 0, 0, 0, 0),
  sampling_end_time = c(0, 0, 0, 0, 0),
  sampling_interval_type = c(0, abs(duration), 0),
  npol_depositing = 1,
  particle_params = c(0, 0, 0),
  dmwsrdre = c(0, 0, 0, 0, 0),
  wrhcicbc = c(0, 0, 0),
  radiactive_decay = 0,
  pol_res = 0,
  control = "CONTROL",
  verbose = FALSE
)
}
\arguments{
\item{df}{data.frame of receptor information. Must include, "year",
"month", "day", "hour" (0:23), "minute", "latitude", "longitude", "altitude"}

\item{year}{year, if missing df.}

\item{month}{month, if missing df.}

\item{day}{day, if missing df.}

\item{hour}{hour, if missing df.}

\item{minute}{minute, if missing df.}

\item{lat}{latitude, if missing df.}

\item{lon}{longitude, if missing df.}

\item{alt}{altitude, if missing df.}

\item{start_time}{String to be added, default derived from df.}

\item{nlocations}{number of locations.}

\item{duration}{number of hours of release. (Negative is backwards in time).}

\item{vertical_motion}{Vertical motion option.  (0:data 1:isob 2:isen 3:dens
4:sigma 5:diverg 6:msl2agl 7:average 8:damped).
. The default "data" selection will use the meteorological model's
vertical velocity fields; other options include isobaric, isentropic,
constant density, constant internal sigma coordinate, computed from the
velocity divergence, a special transformation to correct the vertical
velocities when mapped from quasi-horizontal surfaces (such as relative
 to MSL) to HYSPLIT's internal terrain following sigma coordinate, and a
 special option (7) to spatially average the vertical velocity.
 The averaging distance is automatically computed from the ratio of
 the temporal frequency of the data to the horizontal grid resolution.
 Default 5}

\item{top_model_domain}{altitude above ground level (m).}

\item{met}{meteorological models to be used.}

\item{nmet}{Number of days for the meteorological files. Default is
number of days in duration plus two days.
nmet is the number of simultaneous input meteorological files. For instance,
11 means that for each meteorological grid, 11 files are expected. Usually,
the files are daily. Note that the same number of files are required for
each grid in this approach. Hysplit expects something like 2 11,
which means 2 meteorological grids with 11 files each. The number
2 comes from the length of met files.}

\item{metpath}{paths for each meteorological model output.}

\item{metformat}{format to applied to the meteorological daily file}

\item{ngases}{Default 1.}

\item{gas}{default "Foot".}

\item{emissions_rate}{Default 0}

\item{hour_emissions}{hour release, depend of type of release, instantaneous 0.01,
continuous 1.}

\item{release_start}{derived from df.}

\item{nsim_grids}{Number of simulated grids.}

\item{center_conc_grids}{center coordinates for conc grid.}

\item{grid_spacing}{grid spacing, default is 0.1 degree.}

\item{grid_span}{model extension in degrees by dimension.}

\item{nconc}{name of the concentration file, default is "cdump"}

\item{nvert_levels}{number of vertical levels}

\item{height_vert_levels}{hright vertical levels (50)}

\item{sampling_start_time}{2 digits year, month, day, hour, minute}

\item{sampling_end_time}{2 digits year, month, day, hour, minute}

\item{sampling_interval_type}{type, hour, minure}

\item{npol_depositing}{number of pollutant depositing}

\item{particle_params}{Particle diameter, density, and shape}

\item{dmwsrdre}{DepVel, MW, SurfReRa, DifRa, EHenry}

\item{wrhcicbc}{Wet removal: HC, incloud, belowcloud}

\item{radiactive_decay}{days}

\item{pol_res}{Pollutant Resuspension (1/m)}

\item{control}{name of the file, default "CONTROL"}

\item{verbose}{logical, to show more information}
}
\value{
A CONTROL file
}
\description{
This function creates a CONTROL file for HYSPLIT model.
It uses inputs from a data.frame with the receptor information.
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_read(index)
df <- dt[1]
control_file <- tempfile()
obs_hysplit_control(df, control = control_file)
ff <- readLines(control_file)

cat(ff, sep =  "\n")
}
}
