% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_obs_hysplit.R
\name{obs_hysplit_control_read}
\alias{obs_hysplit_control_read}
\title{obs_hysplit_control_read}
\usage{
obs_hysplit_control_read(control = "CONTROL")
}
\arguments{
\item{control}{CONTROL text file}
}
\value{
A list with CONTROL information
}
\description{
This function creates a CONTROL file for HYSPLIT model.
It uses inputs from a data.frame with the receptor information.
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_read(index)
df <- dt[1]
control_file <- tempfile()
obs_hysplit_control(df, control = control_file)
ff <- readLines(control_file)

cat(ff, sep =  "\n")
obs_hysplit_control_read(control_file)

}
}
