% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_obs_hysplit.R
\name{obs_hysplit_setup}
\alias{obs_hysplit_setup}
\title{obs_hysplit_setup}
\usage{
obs_hysplit_setup(
  idsp = 2,
  capemin = 500,
  vscales = -1,
  kbls = 1,
  kblt = 5,
  kmixd = 0,
  initd = 0,
  veght = 0.5,
  kmix0 = 150,
  numpar = 500,
  maxpar = 500,
  ichem = 8,
  krand = 4,
  varsiwant = c("time", "indx", "lati", "long", "zagl", "zsfc", "foot", "samt", "temp",
    "dswf", "mlht", "dens", "dmas", "sigw", "tlgr"),
  ivmax = length(varsiwant),
  outdt = 15,
  extra_params,
  setup = "SETUP.CFG"
)
}
\arguments{
\item{idsp}{particle dispersion scheme 1:HYSPLIT 2:STILT}

\item{capemin}{-1 no convection; -2 Grell convection scheme;
-3 extreme convection; >0 enhanced vertical mixing when CAPE exceeds this value (J/kg)}

\item{vscales}{vertical Lagrangian time scale (sec) for stable PBL}

\item{kbls}{boundary layer stability derived from 1:fluxes 2:wind_temperature}

\item{kblt}{boundary layer turbulence parameterizations 1:Beljaars 2:Kanthar
3:TKE 4:Measured 5:Hanna}

\item{kmixd}{mixed layer obtained from 0:input 1:temperature 2:TKE 3:modified Richardson}

\item{initd}{initial distribution, particle, puff, or combination}

\item{veght}{Height below which particle's time is spent is tallied to
calculate footprint for PARTICLE_STILT.DAT less than or equal to 1.
0: fraction of PBL height; greater than 1.0: heightAGL (m)}

\item{kmix0}{minimum mixing depth}

\item{numpar}{number of puffs or particles to released per cycle}

\item{maxpar}{maximum number of particles carried in simulation}

\item{ichem}{chemistry conversion modules 0:none 1:matrix 2:convert 3:dust
4: conc grid equal to met grid, 5: divide output mass by air density (kg/m3)
to sum as mixing ration, 7: transport deposited particles on the ocean surface,
8: stilt mode mixing ratio and varying layer,
9: set concentration layer one to a fraction of the boundary layer,
10: restructure concentration grid into time-varying transfer matrix,
11: enable daughter produyct calculation.}

\item{krand}{0 method to calculate random number
0=precompute if NUMPAR greater than 5000 or dynamic if NUMPAR less than or equal to 5000;
1=precalculated; 2=calculated in pardsp;
3=none; 4=random initial seed number and calculated in pardsp;
10=same as 0 with random initial seed for non-dispersion applications;
11=same as 1 with random initial seed for non-dispersion applications;
12=same as 2 with random initial seed for non-dispersion applications;
13=same as 3 with random initial seed for non-dispersion applications}

\item{varsiwant}{='TIME','INDX','LONG','LATI','ZAGL','SIGW','TLGR','ZSFC', TEMP',
'SAMT','FOOT','SHTF','DMAS','DENS','RHFR','SPHU','DSWF','WOUT','MLHT','PRES'
variables written to PARTICLE_STILT.DAT.

However, the default in this case is:
c('time','indx', 'lati','long','zagl', 'zsfc','foot','samt', 'temp',
'dswf','mlht','dens','dmas','sigw','tlgr'}

\item{ivmax}{0 number of variables written to PARTICLE_STILT.DAT. Must
equal the number of variables listed for variable VARSIWANT}

\item{outdt}{Default 15. defines the output frequency in minutes of the endpoint positions
in the PARTICLE.DAT file when the STILT emulation mode is configured.
The default value of 0 results in output each time step while the positive
value gives the output frequency in minutes. A negative value disables output. The
output frequency should be an even multiple of the time step and be evenly
divisible into 60. In STILT emulation mode, the time step is forced to one minute.}

\item{extra_params}{more parameters}

\item{setup}{Default SETUP.CFG}
}
\value{
A SETUP.CFG file
}
\description{
This function creates a SETUP.CFG file for HYSPLIT model.
}
\note{
The var description comes from hysplit 5.3 manual page 214
}
\examples{
{
# Do not run
setup_file <- tempfile()
obs_hysplit_setup(setup = setup_file)
cat(readLines(setup_file),sep =  "\n")
}
}
