% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_id2pos}
\alias{obs_id2pos}
\title{obs_id2pos}
\usage{
obs_id2pos(id, sep = "x", asdf = FALSE)
}
\arguments{
\item{id}{string}

\item{sep}{string}

\item{asdf}{Logical, to return as data.frame or not}
}
\value{
data.frame with time and location based on input
}
\description{
return ndata.frame based on footprint/receptor id
}
\details{
Helpers Legacy
}
\examples{
\dontrun{
# Do not run
id <- '2002x08x03x10x45.00Nx090.00Ex00030'
(obs_id2pos(id, asdf = TRUE) -> dx)
id <- c('2002x08x03x10x00x45.000Nx090.000Ex00030',
        '2002x08x03x10x55x45.335Sx179.884Wx00030')
(obs_id2pos(id) -> dx)
(obs_id2pos(id, asdf = TRUE) -> dx)
(obs_id2pos(rep(id, 2)) -> dx)
(obs_id2pos(rep(id, 2), asdf = TRUE) -> dx)
}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_grid}()},
\code{\link{obs_info2id}()},
\code{\link{obs_julian}()},
\code{\link{obs_normalize_dmass}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
