% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_info2id}
\alias{obs_info2id}
\title{obs_info2id}
\usage{
obs_info2id(yr, mo, dy, hr, mn = 0, lat, lon, alt, sep = "x", long = T)
}
\arguments{
\item{yr}{year}

\item{mo}{month}

\item{dy}{day}

\item{hr}{hour}

\item{mn}{minute}

\item{lat}{latitude}

\item{lon}{longitude}

\item{alt}{altitude above ground level}

\item{sep}{character, default"x"}

\item{long}{Logical, to add minute, and rounded with 2 decimals, instead of 4. default TRUE}
}
\value{
a string with the receptor id
}
\description{
return footprint/receptor id
}
\examples{
{
\dontrun{
# Do not run
obs_info2id(yr = 2002,
            mo = 8,
            dy = 3,
            hr = 10,
            mn = 0,
            lat = 42,
            lon = -90,
            alt = 1) [1]
}
}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_decimal_to_POSIX}()},
\code{\link{obs_id2pos}()},
\code{\link{obs_interpret_udunits_time}()},
\code{\link{obs_julian}()},
\code{\link{obs_load_ncdf}()},
\code{\link{obs_normalize_dmass}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
