% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_interpret_udunits_time}
\alias{obs_interpret_udunits_time}
\title{obs_interpret_udunits_time}
\usage{
obs_interpret_udunits_time(vals, unitstring, tz = "UTC")
}
\arguments{
\item{vals}{time array}

\item{unitstring}{month}

\item{tz}{time zone, "UTC" default}
}
\value{
return list of time vars
}
\description{
return list of time vars
}
\examples{
{
\dontrun{
# Do not run
}}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_decimal_to_POSIX}()},
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()},
\code{\link{obs_julian}()},
\code{\link{obs_load_ncdf}()},
\code{\link{obs_normalize_dmass}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
