% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_inv_files.R
\name{obs_invfiles}
\alias{obs_invfiles}
\title{Generate files to perform inverse modeling}
\usage{
obs_invfiles(
  master,
  nc = "nc",
  path,
  value = "value",
  value_factor = 1e+09,
  Type = "continuous",
  SubType = "tower-insitu",
  Surface_Elev = 99999,
  Model_agl,
  Data_agl,
  Event = 99999,
  Institution,
  Scale,
  outdir
)
}
\arguments{
\item{master}{data.table master file after filtering obspack}

\item{nc}{column name of master which indicate NetCDF footprints}

\item{path}{first part of the full path of the .nc footprint files}

\item{value}{column name with the pollutant}

\item{value_factor}{numeric factor}

\item{Type}{"continuous", "flask" or "hatsflask"}

\item{SubType}{"aircraft-insitu", "tower-insitu", "surface-insitu" and so on}

\item{Surface_Elev}{Site elevation, default 99999}

\item{Model_agl}{Model for agl, if missing is column `altitude_final`}

\item{Data_agl}{Model for agl, if missing is column `altitude_final`}

\item{Event}{Number of event}

\item{Institution}{Institution}

\item{Scale}{Scale}

\item{outdir}{String for the output dir}
}
\value{
A list of data.frame
}
\description{
This function return a list of three elements.
}
\details{
1. Footprints_hera_hysplit with the  full path to the NetCDF footprints\

2. Obs_hysplit concentration associazted iwth each receptor

3. Receptor_info_hysplit with receptor info
}
\examples{
\dontrun{
# do not run
}
}
