% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_itau_to_POSIX}
\alias{obs_itau_to_POSIX}
\alias{itau2date}
\title{obs_itau_to_POSIX}
\usage{
obs_itau_to_POSIX(itaus, year0 = 1999)
}
\arguments{
\item{itaus}{vector of integers representing ONE date or date time}

\item{year0}{default 1999}
}
\value{
return list of time vars
}
\description{
return time vars
}
\examples{
{
\dontrun{
# Do not run
obs_itau_to_POSIX(1)
}}
}
\seealso{
Other helpers legacy time: 
\code{\link{obs_POSIX_to_decimal}()},
\code{\link{obs_POSIX_to_epoch_seconds}()},
\code{\link{obs_POSIX_to_idate}()},
\code{\link{obs_POSIX_to_julian}()},
\code{\link{obs_days_in_month}()},
\code{\link{obs_days_in_year}()},
\code{\link{obs_epoch_seconds_to_POSIX}()},
\code{\link{obs_extract_dates}()},
\code{\link{obs_format_et}()},
\code{\link{obs_idate_to_POSIX}()},
\code{\link{obs_is_leap}()},
\code{\link{obs_julian_to_POSIX}()},
\code{\link{obs_length_POSIXlt}()},
\code{\link{obs_seconds_in_month}()},
\code{\link{obs_seconds_in_year}()},
\code{\link{obs_seq_midmon}()}
}
\concept{helpers legacy time}
