% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_julian}
\alias{obs_julian}
\title{obs_julian}
\usage{
obs_julian(m, d, y, origin.)
}
\arguments{
\item{m}{numeric}

\item{d}{numeric}

\item{y}{numeric}

\item{origin.}{string}
}
\value{
returns day since 1/1/1960
}
\description{
return numeric
}
\details{
Helpers Legacy
}
\examples{
{
\dontrun{
# Do not run
obs_julian(1, 2020, 1)
}
}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_id2pos}()}
}
\concept{helpers legacy}
