% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_list.dt}
\alias{obs_list.dt}
\title{list.dt}
\usage{
obs_list.dt(ldf, na, verbose = TRUE)
}
\arguments{
\item{ldf}{list of data.frames}

\item{na}{common names in the final data.frame}

\item{verbose}{Logical to show more information}
}
\value{
long data.table
}
\description{
Some treatments for list of data.frames
}
\note{
1. Filter out empty data.frames
2. identify common names
3. rbindlist and return data.table
}
\examples{
{
\dontrun{
# Do not run
}
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
