% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_load_ncdf}
\alias{obs_load_ncdf}
\title{obs_load_ncdf}
\usage{
obs_load_ncdf(
  ncname,
  vars = NULL,
  grps = NULL,
  dims = TRUE,
  useoldfoot = FALSE,
  attribs = "Conventions",
  var.attribs = NULL,
  gz = FALSE,
  tmpdir = tempdir(),
  tmpfile = tempfile("load.ncdf4", tmpdir = tmpdir),
  clean = TRUE,
  lowercase = FALSE,
  gsubdot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ncname}{pathname of NetCDF4 file}

\item{vars}{specify variables to retrieve otherwise all are returned
(NOTE: '/' indicates only return root group)}

\item{grps}{groups}

\item{dims}{return variable dimensions TRUE/FALSE}

\item{useoldfoot}{old footprints have dim order: lat,lon,time but new footprints
(CF-1.6 convention and on) have dim order: lon,lat,time; this option flips
lat/lon dim order of returned footprints for backwards compatibility;
TRUE => flips new footprints to old convention;
FALSE => flips old footprints to new convention}

\item{attribs}{array of names of global file attributes to return (NOTE: this is for
backwards compatibility - current version automatically returns all
global file attributes)}

\item{var.attribs}{array of names of variable attributes to return (NOTE: this is for
backwards compatibility - current version automatically returns all
attributes for variables and dimensions)}

\item{gz}{netCDF file is gzipped. Note that .nc may be already compressed.}

\item{tmpdir}{temporary directory to unzip netCDF file}

\item{tmpfile}{temporary unzipped netCDF file name}

\item{clean}{remove temporary file after processing}

\item{lowercase}{return all list names in lowercase}

\item{gsubdot}{substitute '_' with '.' in all list variable names}

\item{verbose}{logical, to show more messages}
}
\value{
list of NetCDF vars and attributes
}
\description{
NetCDF list of var and attributes
}
\examples{
{
\dontrun{
# Do not run
}}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_decimal_to_POSIX}()},
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()},
\code{\link{obs_interpret_udunits_time}()},
\code{\link{obs_julian}()},
\code{\link{obs_normalize_dmass}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
