% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_make_hysplit_nc4}
\alias{obs_make_hysplit_nc4}
\title{obs_make_hysplit_nc4}
\usage{
obs_make_hysplit_nc4(
  ident,
  ncname = paste("hysplit", ident, ".nc", sep = ""),
  origident = FALSE,
  part = FALSE,
  foot = FALSE,
  horizconv = FALSE,
  avghorizconv = FALSE,
  dvmr = FALSE,
  dvmrmean = FALSE,
  dvmrstd = FALSE,
  runinfo = FALSE,
  check = FALSE,
  emit = FALSE,
  init = FALSE,
  avginit = FALSE,
  partdatname = "part",
  footname = NULL,
  horizconvname = NULL,
  avghorizconvname = NULL,
  dvmrname = NULL,
  dvmrmeanname = NULL,
  dvmrstdname = NULL,
  initname = NULL,
  avginitname = NULL,
  horizconvdatunits = NULL,
  avghorizconvdatunits = NULL,
  dvmrdatunits = NULL,
  dvmrmeandatunits = NULL,
  dvmrstddatunits = NULL,
  initdatunits = NULL,
  avginitdatunits = NULL,
  origidentdat = NULL,
  origutctime.format = NULL,
  partdat = NULL,
  footdat = NULL,
  horizconvdat = NULL,
  avghorizconvdat = NULL,
  dvmrdat = NULL,
  dvmrmeandat = NULL,
  dvmrstddat = NULL,
  runinfodat = NULL,
  checkdat = NULL,
  emitdat = NULL,
  initdat = NULL,
  avginitdat = NULL,
  partpath = NULL,
  checkpath = partpath,
  footpath = paste(partpath, "footprints/", sep = ""),
  horizconvpath = paste(partpath, "convolve/", sep = ""),
  avghorizconvpath = paste(partpath, "convolve/", sep = ""),
  initpath = paste(partpath, "boundary/", sep = ""),
  avginitpath = paste(partpath, "boundary/", sep = ""),
  runinfopath = NULL,
  targetdir,
  errorlog = paste("ncerrors", ident, ".txt", sep = ""),
  appendnc = FALSE,
  plotfoot = FALSE,
  plotmeantrajec = FALSE,
  sourcepath = NULL,
  charmissval = NULL,
  global.att.list = NULL,
  global.att.files = NULL,
  ident.start = NULL
)
}
\arguments{
\item{ident}{identifier}

\item{ncname}{netcdf name file format hysplitid.nc}

\item{origident}{logical, default FALSE, include most accurate receptor agl,lat,lon, utctime fields,  before rounding}

\item{part}{logical, default FALSE, include particle array}

\item{foot}{logical, default FALSE, include foot}

\item{horizconv}{logical, default FALSE, include results from horizontally convolving footprints with
fluxes for each particle and summing particle over time}

\item{avghorizconv}{logical, default FALSE, include results from horizontally convolving footprints with
fluxes for each particle and summing particle over time, avg particles}

\item{dvmr}{logical, default FALSE, include delta volume mixing ratio results from convolve-fast.r
(replaces horizconv and avghorizconv functionality)}

\item{dvmrmean}{logical, default FALSE, include climatological mean of delta volume mixing ratio results}

\item{dvmrstd}{logical, default FALSE, include climatological standard dev of volume mixing ratio results}

\item{runinfo}{logical, default FALSE, text summary of run parameters}

\item{check}{logical, default FALSE, output from Trajeccheck}

\item{emit}{logical, default FALSE, emission window description}

\item{init}{logical, default FALSE, include init mixing ratio values}

\item{avginit}{logical, default FALSE, include avginit mixing ratio value}

\item{partdatname}{string, default 'part', name of particle table, with the following convention:
part       => whole particle table  (default)
partfoot   => particle table thinned to retain rows where foot > 0
part3d     => particle table thinned to retain rows every nhrs
endpts     => particle table thinned to retain rows with trajectory endpts}

\item{footname}{default NULL, name of footprint to distinguish among multiple footprints -e.g. foot1, foot2, etc}

\item{horizconvname}{name of horizontal convolve mixing ratio values (should reflect species nd source -
e.g. co2horizconvCT, co2horizconvvul, etc)}

\item{avghorizconvname}{name of avg horizontal convolve mixing ratio (should reflect species and source -
e.g. co2avghorizconvCT, co2avghorizconvvul, etc)}

\item{dvmrname}{name of delta volume mixing ratio results (should reflect species  and source -
e.g. co2dvmrCT, co2dvmrvul, etc)}

\item{dvmrmeanname}{name of climatological mean of delta volume mixing ratio results (should
reflect species and source - e.g. co2dvmrmeanCT, co2dvmrmeanvul, etc)}

\item{dvmrstdname}{name of climatological std of delta volume mixing ratio results (should reflect
species and source - e.g. co2dvmrstdCT, co2dvmrstdvul, etc)}

\item{initname}{name of init mixing ratio values (should reflect species and source) (e.g.
co2initCT, co2initGV, coinitGV, etc)}

\item{avginitname}{name of average init mixing ratio (should reflect species and source)
(e.g. co2avginitCT, co2avginitGV, coavginitGV, etc)}

\item{horizconvdatunits}{default NULL, string indicating assigned units to horizconvdat}

\item{avghorizconvdatunits}{default NULL, string indicating assigned units to avghorizconvdat}

\item{dvmrdatunits}{default NULL, string indicating assigned units to dvmrdat}

\item{dvmrmeandatunits}{default NULL, string indicating assigned units to dvmrmeandat}

\item{dvmrstddatunits}{default NULL, string indicating assigned units to dvmrstddat}

\item{initdatunits}{default NULL, string indicating assigned units to initdat}

\item{avginitdatunits}{default NULL, string indicating assigned units to avginitdat}

\item{origidentdat}{default NULL, list of original receptor agl, lat, lon, utctime fields}

\item{origutctime.format}{default NULL, string format of origutctime from receptor list e.g. '%Y-%m-%dT%H:%M:%S'}

\item{partdat}{default NULL,  passed in R object (use getr by default)}

\item{footdat}{default NULL,  passed in R object (use getr by default)}

\item{horizconvdat}{default NULL,  passed in R object (use getr by default)}

\item{avghorizconvdat}{default NULL,  passed in R object (use getr by default)}

\item{dvmrdat}{default NULL, matrix of dvmr results}

\item{dvmrmeandat}{default NULL, matrix of dvmrmean results}

\item{dvmrstddat}{default NULL, matrix of dvmrstd results}

\item{runinfodat}{default NULL, passed in R object (use getr by default)}

\item{checkdat}{default NULL, passed in R object (use getr by default)}

\item{emitdat}{default NULL, data frame/2D array}

\item{initdat}{default NULL, passed in R object (use getr by default)}

\item{avginitdat}{default NULL, passed in R object (use getr by default)}

\item{partpath}{default NULL, location of particle file}

\item{checkpath}{default partpath, location of particle file}

\item{footpath}{paste(partpath, "footprints/", sep=""), location of footprint file}

\item{horizconvpath}{paste(partpath, "convolve/", sep="") location of horizontal convolve summarized results file}

\item{avghorizconvpath}{paste(partpath, "convolve/", sep="") location of horizontal convolve summarized results file}

\item{initpath}{paste(partpath, "boundary/", sep="") location of init file}

\item{avginitpath}{paste(partpath, "boundary/", sep="") location of init file}

\item{runinfopath}{default NULL,  location of runinfo file}

\item{targetdir}{destination directory}

\item{errorlog}{paste("ncerrors",ident,".txt",sep=""), name of errorlog file; use NA to suppress}

\item{appendnc}{default FALSE, to append to nc}

\item{plotfoot}{default FALSE, to plot footprints}

\item{plotmeantrajec}{default FALSE, to plot mean trajectory}

\item{sourcepath}{default NULL}

\item{charmissval}{default NULL, missing value for character variables (ncvar_def requires NA or NULL depending on the version of ncdf4)}

\item{global.att.list}{default NULL, list of global attributes (only used if !appendnc)}

\item{global.att.files}{default NULL, list of filenames whose contents are turned into global attributes
(only used if !appendnc and global.att.list is empty)}

\item{ident.start}{default NULL,  time related to hysplit start time (may different from nearest hour to ident if partperhymodelc was turned on)}
}
\value{
return dates
}
\description{
makes nc
}
\note{
Convention is that the Julian date of the epoch is 1.0 (not 0!)
}
\examples{
{
\dontrun{
# Do not run
}}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_decimal_to_POSIX}()},
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()},
\code{\link{obs_interpret_udunits_time}()},
\code{\link{obs_julian}()},
\code{\link{obs_load_ncdf}()},
\code{\link{obs_normalize_dmass}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
