% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_meta_obs.R
\name{obs_meta}
\alias{obs_meta}
\title{Read obspack metadata}
\usage{
obs_meta(
  index,
  verbose = TRUE,
  n_site_code = 15,
  n_site_name = 15,
  n_site_latitude = 18,
  n_site_longitude = 19,
  n_site_country = 18,
  n_dataset_project = 21,
  n_lab = 16,
  n_scales = 31,
  n_site_elevation = 20,
  n_altitude_comment = 22,
  n_utc = 18,
  fill_value = -1e+34,
  as_list = FALSE
)
}
\arguments{
\item{index}{data.table}

\item{verbose}{Logical to show more information}

\item{n_site_code}{number of characters extraced from metadata after search}

\item{n_site_name}{number of characters extraced from metadata after search}

\item{n_site_latitude}{number of characters extraced from metadata after search}

\item{n_site_longitude}{number of characters extraced from metadata after search}

\item{n_site_country}{number of characters extraced from metadata after search}

\item{n_dataset_project}{number of characters extraced from metadata after search}

\item{n_lab}{number of characters extraced from metadata after search}

\item{n_scales}{number of characters extraced from metadata after search}

\item{n_site_elevation}{number of characters extraced from metadata after search}

\item{n_altitude_comment}{number of characters extraced from metadata after search}

\item{n_utc}{number of characters extraced from metadata after search}

\item{fill_value}{fill value. Appeared in aoa_aircraft-flask_19_allvalid.txt}

\item{as_list}{Logical to return as list}
}
\value{
A data.frame with with an index obspack.
}
\description{
Read obspack metadata
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_meta(index)
}
}
