% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_obs_nc.R
\name{obs_nc}
\alias{obs_nc}
\title{obs_nc}
\usage{
obs_nc(
  lat,
  lon,
  time_nc,
  vars_out = c("total", "bio", "ocn", "fossil", "fire"),
  units_out = "(ppb/nanomol m-2 s-1)*nanomol m-2 s-1",
  nc_out,
  larrays,
  verbose = FALSE
)
}
\arguments{
\item{lat}{lats array}

\item{lon}{longs arrau}

\item{time_nc}{Times.}

\item{vars_out}{names for the variables to be created in the NetCDF.}

\item{units_out}{units for the NetCDF variables to be created.}

\item{nc_out}{path for the created NetCDF.}

\item{larrays}{list of arrays, length equal to vars_out.}

\item{verbose}{Logical, to display more information.}
}
\value{
A NetCDF file with the custom attributes and units
}
\description{
Creates NetCDF based on dimension from another NetCDF
with custom dimensions, and attributes
}
\examples{
\dontrun{
# nc_path <- paste0("Z:/footprints/aircraft/flask/2018",
# "/04/hysplit2018x04x08x15x15x38.7459Nx077.5584Wx00594.nc")
# foot <- obs_nc_get(nc_path, all = TRUE)
# nco <- paste0(tempfile(), "2.nc")
# file.remove(nco)
# obs_nc(lat = foot$lat,
#        lon = foot$lon,
#        time_nc = ISOdatetime(2018, 4, 8, 15, 15, 38),
#        vars_out = c("a", "b"),
#        nc_out  = nco,
#        larrays = list(a = foot, b = foot),
#        verbose = TRUE)
}
}
