% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_normalize_dmass}
\alias{obs_normalize_dmass}
\title{obs_normalize_dmass}
\usage{
obs_normalize_dmass(part = NULL)
}
\arguments{
\item{part}{particle data.table}
}
\value{
return footprint
}
\description{
add columns of
}
\examples{
{
\dontrun{
# Do not run
}}
#' # Do not run
}
\seealso{
Other helpers legacy: 
\code{\link{obs_grid}()},
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()},
\code{\link{obs_julian}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
