% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_out}
\alias{obs_out}
\title{outersect}
\usage{
obs_out(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
vector opposite of intersect
}
\description{
Just the opposite of intersect
}
\details{
Helpers
}
\examples{
{
\dontrun{
# Do not run
}
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{robspack-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
