% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_plot.R
\name{obs_plot}
\alias{obs_plot}
\title{Read obspack metadata}
\usage{
obs_plot(
  dt,
  time,
  y = "value",
  yfactor = 1,
  colu = "site_code",
  type = "p",
  n = if (length(unique(dt[[colu]])) == 1) unique(dt[[colu]]) else
    unique(dt[[colu]])[1:2],
  pal = cptcity::find_cpt("qual")[6],
  verbose = TRUE,
  xlab = "time",
  ylab = "value",
  xlim = range(dt[[time]], na.rm = TRUE),
  ylim = range(dt[[y]], na.rm = TRUE),
  ...
)
}
\arguments{
\item{dt}{data.table}

\item{time}{x axis column (time)}

\item{y}{y axis column, default "value"}

\item{yfactor}{factor of y}

\item{colu}{column to plot by color, default site_code}

\item{type}{type of plot, default "p"}

\item{n}{Character indicating `colu` to subset, for instance,
if you want to plot "site_code", here include all the
site_code that you want to plot}

\item{pal}{Color palette name see \link[cptcity]{cpt},
default "cb_qual_Accent_08"}

\item{verbose}{Logical to show more information}

\item{xlab}{Character, xlab}

\item{ylab}{Character, ylab}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{...}{plot arguments}
}
\value{
Plot
}
\description{
Read obspack metadata
}
\examples{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_read(index)
obs_plot(dt, time = "time")
}
}
