% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_POSIX_to_julian}
\alias{obs_POSIX_to_julian}
\title{obs_POSIX_to_julian}
\usage{
obs_POSIX_to_julian(times, epoch = ISOdatetime(1, 1, 1, 0, 0, 0, tz = "UTC"))
}
\arguments{
\item{times}{time}

\item{epoch}{default 0001-01-01}
}
\value{
return dates
}
\description{
return time
}
\note{
Convention is that the Julian date of the epoch is 1.0 (not 0!)
}
\examples{
{
\dontrun{
# Do not run
}}
}
\seealso{
Other helpers legacy time: 
\code{\link{obs_POSIX_to_decimal}()},
\code{\link{obs_POSIX_to_epoch_seconds}()},
\code{\link{obs_POSIX_to_idate}()},
\code{\link{obs_days_in_month}()},
\code{\link{obs_days_in_year}()},
\code{\link{obs_epoch_seconds_to_POSIX}()},
\code{\link{obs_extract_dates}()},
\code{\link{obs_format_et}()},
\code{\link{obs_idate_to_POSIX}()},
\code{\link{obs_is_leap}()},
\code{\link{obs_itau_to_POSIX}()},
\code{\link{obs_julian_to_POSIX}()},
\code{\link{obs_length_POSIXlt}()},
\code{\link{obs_seconds_in_month}()},
\code{\link{obs_seconds_in_year}()},
\code{\link{obs_seq_midmon}()}
}
\concept{helpers legacy time}
