% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_rbind}
\alias{obs_rbind}
\title{rbind obspack}
\usage{
obs_rbind(dt1, dt2, verbose = TRUE)
}
\arguments{
\item{dt1}{first data.table}

\item{dt2}{second data.table}

\item{verbose}{logical to show more information}
}
\value{
numeric vector
}
\description{
return rbind obs data.tables
}
\examples{
\dontrun{
# Do not run
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
