% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_read_csvy}
\alias{obs_read_csvy}
\title{reads CSVY}
\usage{
obs_read_csvy(f, n = 100, ...)
}
\arguments{
\item{f}{path to csvy file}

\item{n}{number of files to search for "---" yaml}

\item{...}{extra data.table arguments}
}
\description{
Reads CSVY, print YAML and fread file.
}
\examples{
{
df <- data.frame(a = rnorm(n = 10),
                 time = Sys.time() + 1:10)

f <- paste0(tempfile(), ".csvy")
notes <- c("notes",
           "more notes")
obs_write_csvy(dt = df, notes = notes, out = f)
s <- obs_read_csvy(f)
s
# or
readLines(f)
data.table::fread(f)
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{robspack-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
