% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_roundtime}
\alias{obs_roundtime}
\title{round seconds from "POSIXct" "POSIXt"  classes}
\usage{
obs_roundtime(x, n = 10)
}
\arguments{
\item{x}{time as "POSIXct" "POSIXt"}

\item{n}{factor}
}
\value{
numeric vector
}
\description{
return rounded seconds from time
}
\examples{
\dontrun{
# Do not run
x <- Sys.time() + seq(1, 55, 1)
paste0(x,"  ",
       obs_roundtime(x), "  ",
       obs_freq(data.table::second(x),
                seq(0, 55, 10)))
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
