% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_seconds_in_month}
\alias{obs_seconds_in_month}
\title{obs_seconds_in_month}
\usage{
obs_seconds_in_month(year, month)
}
\arguments{
\item{year}{year}

\item{month}{month (1 to 12)}
}
\value{
return dates
}
\description{
return time
}
\examples{
{
\dontrun{
# Do not run
}}
}
\seealso{
Other helpers legacy time: 
\code{\link{obs_POSIX_to_decimal}()},
\code{\link{obs_POSIX_to_epoch_seconds}()},
\code{\link{obs_POSIX_to_idate}()},
\code{\link{obs_POSIX_to_julian}()},
\code{\link{obs_days_in_month}()},
\code{\link{obs_days_in_year}()},
\code{\link{obs_epoch_seconds_to_POSIX}()},
\code{\link{obs_extract_dates}()},
\code{\link{obs_format_et}()},
\code{\link{obs_idate_to_POSIX}()},
\code{\link{obs_is_leap}()},
\code{\link{obs_itau_to_POSIX}()},
\code{\link{obs_julian_to_POSIX}()},
\code{\link{obs_length_POSIXlt}()},
\code{\link{obs_seconds_in_year}()},
\code{\link{obs_seq_midmon}()}
}
\concept{helpers legacy time}
