% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate_obs.R
\name{obs_select_sec}
\alias{obs_select_sec}
\title{Select Observations by closest time (seconds)}
\usage{
obs_select_sec(
  dt,
  origin = "1970-01-01",
  tz = "UTC",
  seconds = 30,
  verbose = TRUE
)
}
\arguments{
\item{dt}{data.table}

\item{origin}{a date-time object, default "1970-01-01"}

\item{tz}{time zone string}

\item{seconds}{seocnds target to select and filter data}

\item{verbose}{logical to show more information}
}
\value{
A filtered data.frame.
}
\description{
This function select by closest time.
Good for aircrafts
}
\note{
By default add column timeUTC based on the input column `key_time`
}
\examples{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_read(index)
dx <- obs_read(index, expr = "altitude_final == '5800'")
dx <- obs_addtime(dx[, -"site_code"])
dy <- obs_select_sec(dx)
}
}
