% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_index_obs.R, R/01_summary_obs.R
\name{obs_index}
\alias{obs_index}
\alias{obs_summary}
\title{Summary of the ObsPack files}
\usage{
obs_index(
  obs,
  categories = c("aircraft-pfp", "aircraft-insitu", "surface-insitu", "tower-insitu",
    "aircore", "surface-pfp", "shipboard-insitu", "flask"),
  lnchar = 11,
  out = paste0(tempfile(), "_index.csv"),
  verbose = TRUE
)

obs_summary(
  obs,
  categories = c("aircraft-pfp", "aircraft-insitu", "surface-insitu", "tower-insitu",
    "aircore", "surface-pfp", "shipboard-insitu", "flask"),
  lnchar = 11,
  out = paste0(tempfile(), "_index.csv"),
  verbose = TRUE
)
}
\arguments{
\item{obs}{Path to the Obspack GLOBALview txt data}

\item{categories}{character; default are c("aircraft-pfp", "aircraft-insitu",
"surface-insitu", "tower-insitu", "aircore", "surface-pfp", "shipboard-insitu",
"flask"). The idea is that,as the file names include these words,
this function identifies which files has these words and add them as columns.}

\item{lnchar}{Integer; last nchar, default = 11.}

\item{out}{Path to the ObsPack index output}

\item{verbose}{Logical to show more information}
}
\value{
A data.frame with with an index of the obspack Globalview.

A data.frame with with an index of the obspack Globalview.
}
\description{
This function returns a data.frame index of all files
available in obspack.

This function returns a data.frame index of all files
available in ObsPack.
}
\examples{
{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "robspack")
index <- obs_summary(obs)
}
}
{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "robspack")
index <- obs_summary(obs)
}
}
}
