% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_obs_table.R
\name{obs_table}
\alias{obs_table}
\title{Obspack Table Summary}
\usage{
obs_table(
  df,
  cols = c("value", "time", "time_decimal", "latitude", "longitude")
)
}
\arguments{
\item{df}{data.table}

\item{cols}{String of columns to be summarized.}
}
\value{
A data.frame with with an index obspack.
}
\description{
This function reads the obsPack directory
providing a summary for the columns: "value", "time",
"time_decimal", "latitude" and "longitude". These summary
are made by the columns "name", "sector", "site_name",
"site_country", "type_altitude", "lab_1_abbr" and
"site_utc2lst"
}
\examples{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_read(index)
dx <- obs_table(dt)
}
}
