% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_trunc}
\alias{obs_trunc}
\title{Trunc numbers with a desired number of decimals}
\usage{
obs_trunc(n, dec)
}
\arguments{
\item{n}{Numeric  number}

\item{dec}{Integer, number of decimals}
}
\description{
Trunc numbers with a specified number of decimals.
}
\note{
source https://stackoverflow.com/a/47015304/2418532
}
\examples{
{
# in bash:
# printf "\%07.4f" 72.05785
# results in 72.0578
# but:
formatC(72.05785, digits = 4, width = 8, format = "f", flag = "0")
# results in
"072.0579"
# the goal is to obtain the same trunc number as using bash, then:
formatC(obs_trunc(72.05785, 4),
        digits = 4,
        width = 8,
        format = "f",
        flag = "0")
}
}
\seealso{
Other helpers: 
\code{\link{obs_addltime}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_write_csvy}()},
\code{\link{robspack-deprecated}}
}
\concept{helpers}
