% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_write_csvy}
\alias{obs_write_csvy}
\title{Generates YAML and write data.frame}
\usage{
obs_write_csvy(
  dt,
  notes,
  out = paste0(tempfile(), ".csvy"),
  sep = ",",
  nchar.max = 80,
  ...
)
}
\arguments{
\item{dt}{data.table}

\item{notes}{notes.}

\item{out}{outfile path.}

\item{sep}{The separator between columns. Default is ",".}

\item{nchar.max}{Max nchar for str.}

\item{...}{extra data.table arguments}
}
\value{
a csvy file to disk
}
\description{
Add YAML header info and writes a the data.frame
into disk. The YAML section includes notes and str(dt).
}
\examples{
\dontrun{
# Do not run
df <- data.frame(a = rnorm(n = 10),
                 time = Sys.time() + 1:10)

f <- paste0(tempfile(), ".csvy")
notes <- c("notes",
           "more notes")
obs_write_csvy(dt = df, notes = notes, out = f)
readLines(f)
data.table::fread(f, h = TRUE)
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
