% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/robspack-deprecated.R
\name{robspack-deprecated}
\alias{robspack-deprecated}
\title{Number to character}
\arguments{
\item{x}{integer}

\item{char}{Character to be added ad the beginning of x}
}
\value{
transform integer in character
}
\description{
transform integer to character and add zero at left

The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\seealso{
\code{\link{robspack-deprecated}}

Other helpers: 
\code{\link{obs_addltime}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()}
}
\concept{helpers}
\keyword{internal}
