% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robspack-package.R
\docType{package}
\name{robspack-package}
\alias{robspack}
\alias{robspack-package}
\title{robspack: 'robspack' Obspack GLOBALView Filter}
\description{
Reads the OBspack GLOBALView data from NOAA and perform different filters and analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ibarraespinosa/robspack}
  \item \url{https://ibarraespinosa.github.io/robspack/}
  \item Report bugs at \url{https://github.com/ibarraespinosa/robspack/issues/}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarra-espinosa@noaa.gov} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

Other contributors:
\itemize{
  \item Lei Hu \email{lei.hu@noaa.gov} (\href{https://orcid.org/0000-0002-5624-3832}{ORCID}) [contributor]
}

}
\keyword{internal}
