% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorf-package.R
\docType{package}
\name{rtorf-package}
\alias{rtorf}
\alias{rtorf-package}
\title{rtorf: R Tools for Observations, Footprints and Receptors (rtorf) and HYSPLIT Configuration Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a suite of tools to ingest, process, and visualize atmospheric observation data, specifically tailored for the NOAA Global Monitoring Laboratory (GML) ObsPack GLOBALView+ products \url{https://gml.noaa.gov/ccgg/obspack/index.html}. Includes specialized routines for filtering trace gas time-series data based on meteorological and quality control parameters. Additionally, streamlines the generation of configuration files for the Hybrid Single-Particle Lagrangian Integrated Trajectory (HYSPLIT) model \url{https://www.ready.noaa.gov/HYSPLIT.php}, supporting high-throughput parallel processing of trajectory simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noaa-gml/rtorf}
  \item \url{https://noaa-gml.github.io/rtorf/}
  \item Report bugs at \url{https://github.com/noaa-gml/rtorf/issues/}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{zergioibarra@gmail.com} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

Other contributors:
\itemize{
  \item Lei Hu \email{lei.hu@noaa.gov} (\href{https://orcid.org/0000-0002-5624-3832}{ORCID}) [contributor]
}

}
\keyword{internal}
