% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sr}
\alias{sr}
\title{Extacts n last characters}
\usage{
sr(x, n)
}
\arguments{
\item{x}{a character vector.}

\item{n}{integer.}
}
\description{
file extension
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}}
}
\concept{helpers}
