% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_addltime}
\alias{obs_addltime}
\title{local hour (bsed on longitude and time)}
\usage{
obs_addltime(
  dt,
  timeUTC = "timeUTC",
  utc2lt = "site_utc2lst",
  longitude = "longitude"
)
}
\arguments{
\item{dt}{data.table}

\item{timeUTC}{Character indicating the Time column as POSIXct}

\item{utc2lt}{Character indicating the integer column to convert to local time
if available}

\item{longitude}{Character indicating the column with the lingitude}
}
\value{
data.table with local time columns
}
\description{
Calculate an approximation of local hour
}
\note{
time depending n longitude is by John Miller (GML/NOAA)
}
\examples{
{
\dontrun{
# Do not run
}
}
}
\seealso{
Other helpers: 
\code{\link{obs_footname}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{robspack-deprecated}}
}
\concept{helpers}
