% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_addtime_obs.R
\name{obs_addtime}
\alias{obs_addtime}
\title{Add times}
\usage{
obs_addtime(dt, verbose = TRUE, tz = "UTC", timeonly = FALSE)
}
\arguments{
\item{dt}{data.table}

\item{verbose}{Logical, to show more info}

\item{tz}{Timezone, default "UTC"}

\item{timeonly}{return only timeUTC column}
}
\value{
A data.frame with with an index obspack.
}
\description{
This function add timeUTC as POSIX class,
local time and ending sampling time
}
\note{
timeUTC is calculated based on the field column start_time,
the timeUTC_end is calculated using this approach:
1. If the column time_interval is not found, proceed
with the calculation using midpoint_time
2. Else, use column time_interval
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <-  obs_summary(obs)
dt <- obs_read(index)
dt <- obs_addtime(dt)
}
}
\seealso{
Other time: 
\code{\link{obs_addltime}()},
\code{\link{obs_addmtime}()},
\code{\link{obs_addstime}()}
}
\concept{time}
