% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate_obs.R
\name{obs_agg}
\alias{obs_agg}
\title{Aggregates ObsPack by time}
\usage{
obs_agg(
  dt,
  gby = "mean",
  cols = c("year", "month", "day", "hour", "minute", "second", "time", "time_decimal",
    "value", "latitude", "longitude", "altitude", "pressure", "u", "v", "temperature",
    "type_altitude", "altitude_final"),
  time = c("UTC", "solar", "local"),
  byalt = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dt}{obspack data.table}

\item{gby}{(groupby) Character indicating the function to group the
data. Default is "mean" meaning that all numeric variables
will be averaged by the new_second column established in the `idcol`}

\item{cols}{Character which defines columns to be aggregated}

\item{time}{Character to aggregate time (solar, local, UTC)}

\item{byalt}{Logical, to aggregate by altitude or not (used in tower)}

\item{verbose}{logical to show more information}
}
\value{
A data.frame with with an index obspack.
}
\description{
This function add timeUTC as POSIX class,
local time and ending sampling time
}
\note{
By default add column timeUTC with parameters from time loca, solar or UTC
}
\examples{
\dontrun{
# Do not run
dt <- obs_addtime(dt)
}
}
