% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_obs_convolve.R
\name{obs_convolve}
\alias{obs_convolve}
\title{obs_convolve}
\usage{
obs_convolve(
  foot_path = "AAA",
  name_foot = "foot1",
  flon = "foot1lon",
  flat = "foot1lat",
  time_foot,
  flux = "CTCO2",
  df_flux,
  flux_format = "\%Y\%m\%d.nc",
  factor = 1e+09,
  fn = NULL,
  as_list = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{foot_path}{path for footprint (length 1).}

\item{name_foot}{name of the footprint variable in the NetCDF file.}

\item{flon}{name of lons}

\item{flat}{name of lats}

\item{time_foot}{Time of the footprints (in the name file) or third dimension}

\item{flux}{String, default NOAA "CTCO2". Implies bio, ocn, fossil, fire.}

\item{df_flux}{data.table with columns f (full path) and nf file name 'Ymd.nc'}

\item{flux_format}{string with date format 'Ymd.nc'}

\item{factor}{number to multiply fluxes.}

\item{fn}{string with function to aggregate convolved fluxes, e.g. `mean`, `sum`, `max`, etc.}

\item{as_list}{Logical, to retunr list of arrays instead of writting to NetCDF}

\item{verbose}{Logical, to display more information}
}
\description{
This function returns a arrays (or list of arrays) of convolved footpritns with flux
}
\note{
main assumption is that fluxes have same spatial dimensions as footprints
}
\examples{
{
# obs_convolve(...)
}
}
