% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{obs_footname}
\alias{obs_footname}
\title{Expected footprint name}
\usage{
obs_footname(year, month, day, hour, minute, lat, lon, alt, out, ...)
}
\arguments{
\item{year}{numeric number}

\item{month}{numeric number}

\item{day}{numeric number}

\item{hour}{numeric number}

\item{minute}{numeric number}

\item{lat}{numeric number}

\item{lon}{numeric number}

\item{alt}{numeric number}

\item{out}{outfile path.}

\item{...}{data.table::fwrite arguments.}
}
\description{
return the expected name for the NetCDF footprint
}
\note{
source https://stackoverflow.com/a/47015304/2418532

# IMPORTANT!!!
# This function will generate the expected NetCDF file name.
# It assumes that the name was generated under the following considerations:
# time variables (year, month, day, etc) have a format of two digits, eg "0.1"
# latitude and longitude have been round with 4 decimals
# The format for latitude is 2 integers, a point and 4 decimals
# The format for longitude is 3 integers, a point and 4 decimals

# In other words, it is similar to `obs_format`, but `obs_footname`,
generates the expected name.
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_format}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
