% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorf-deprecated.R
\name{obs_format}
\alias{obs_format}
\title{Formatting data}
\usage{
obs_format()
}
\value{
formatted characters
}
\description{
Formatting data
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
Other helpers: 
\code{\link{fex}()},
\code{\link{obs_footname}()},
\code{\link{obs_freq}()},
\code{\link{obs_list.dt}()},
\code{\link{obs_out}()},
\code{\link{obs_rbind}()},
\code{\link{obs_read_csvy}()},
\code{\link{obs_roundtime}()},
\code{\link{obs_trunc}()},
\code{\link{obs_write_csvy}()},
\code{\link{rtorf-deprecated}},
\code{\link{sr}()}
}
\concept{helpers}
