% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_grid}
\alias{obs_grid}
\title{obs_grid}
\usage{
obs_grid(min.x, max.x, min.y, max.y, numpix.x, numpix.y, coarse.factor = 1)
}
\arguments{
\item{min.x}{grid indices}

\item{max.x}{grid indices}

\item{min.y}{grid indices}

\item{max.y}{grid indices}

\item{numpix.x}{number pixels x}

\item{numpix.y}{number pixels y}

\item{coarse.factor}{integer to coarse the grid resolution}
}
\value{
return footprint
}
\description{
add columns of
}
\seealso{
Other helpers legacy: 
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()},
\code{\link{obs_julian}()},
\code{\link{obs_normalize_dmass}()},
\code{\link{obs_traj_foot}()}
}
\concept{helpers legacy}
