% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_legacy.R
\name{obs_julian}
\alias{obs_julian}
\title{obs_julian}
\usage{
obs_julian(y, m, d, origin., legacy = FALSE, verbose = TRUE)

obs_julian(y, m, d, origin., legacy = FALSE, verbose = TRUE)
}
\arguments{
\item{y}{year}

\item{m}{month}

\item{d}{day}

\item{origin.}{string}

\item{legacy}{logical, to use legacy code}

\item{verbose}{logical, to show more messages}

\item{hr}{hour}
}
\value{
returns day since 1/1/1960

a string with the receptor id
}
\description{
return numeric

return footprint/receptor id
}
\details{
Helpers Legacy
}
\examples{
{
\dontrun{
# Do not run
obs_julian(1, 2020, 1)
}
}
{
\dontrun{
# Do not run
obs_julian(y = 2002,
            m = 8,
            d = 3,
            legacy = TRUE)
obs_julian(y = 2002,
            m = 8,
            d = 3,
            legacy = FALSE)
}
}
}
\seealso{
Other helpers legacy: 
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()}

Other helpers legacy: 
\code{\link{obs_id2pos}()},
\code{\link{obs_info2id}()}
}
\concept{helpers legacy}
