% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_obs_nc.R
\name{obs_nc_get}
\alias{obs_nc_get}
\title{obs_nc_get}
\usage{
obs_nc_get(
  nc_path = "AAA",
  nc_name = "foot1",
  nc_lat = "foot1lat",
  nc_lon = "foot1lon",
  verbose = FALSE
)
}
\arguments{
\item{nc_path}{String pointing to the target NetCDF}

\item{nc_name}{String indicating the spatial array}

\item{nc_lat}{String to extract latitude.}

\item{nc_lon}{String to extract longitude}

\item{verbose}{Logical, to display more information.}
}
\value{
A NetCDF file with the custom attributes and units
}
\description{
Reads NetCDF var and returns the spatial array
}
\examples{
{
# Do not run
#nc_path <- paste0("Z:/footprints/aircraft/flask/2018/04",
#"/hysplit2018x04x08x15x15x38.7459Nx077.5584Wx00594.nc")
#f <- obs_nc_get(nc_path = nc_path)
}
}
