% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_obs.R
\name{obs_read_nc}
\alias{obs_read_nc}
\title{Read obspack (.nc)}
\usage{
obs_read_nc(
  index,
  categories = "flask",
  att = FALSE,
  expr = NULL,
  solar_time = TRUE,
  as_list = FALSE,
  verbose = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{index}{data.table}

\item{categories}{character; ONE category
: of c("aircraft-pfp", "aircraft-insitu",
"surface-insitu", "tower-insitu", "aircore", "surface-pfp", "shipboard-insitu",
"flask").}

\item{att}{Logical, if global attributes should be added to data.table}

\item{expr}{String expressions to filter data.table internally}

\item{solar_time}{Logical, add solar time?}

\item{as_list}{Logical to return as list}

\item{verbose}{Logical to show more information}

\item{warnings}{Logical to show warnings when reading NetCDF, especially
global attributes}
}
\value{
A data.frame with with an index obspack.
}
\description{
Each obspack file has a header with metadata and this
function reads selected fields from the metadata and add
them as columns. This new columns are used later to
be filtered
}
\examples{
{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
dt <- obs_read(index)
}
}
