% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_summary_obs.R, R/rtorf-deprecated.R
\name{obs_summary}
\alias{obs_summary}
\alias{obs_index}
\title{Summary of the ObsPack files (.txt)}
\usage{
obs_summary(
  obs,
  categories = c("aircraft-pfp", "aircraft-insitu", "surface-insitu", "tower-insitu",
    "aircore", "surface-pfp", "shipboard-insitu", "flask"),
  lnchar = 11,
  out = paste0(tempfile(), "_index.csv"),
  verbose = TRUE,
  aslist = FALSE
)

obs_index(
  obs,
  categories = c("aircraft-pfp", "aircraft-insitu", "surface-insitu", "tower-insitu",
    "aircore", "surface-pfp", "shipboard-insitu", "flask"),
  lnchar = 11,
  out = paste0(tempfile(), "_index.csv"),
  verbose = TRUE
)
}
\arguments{
\item{obs}{Path to the Obspack GLOBALview txt data}

\item{categories}{character; default are c("aircraft-pfp", "aircraft-insitu",
"surface-insitu", "tower-insitu", "aircore", "surface-pfp", "shipboard-insitu",
"flask"). The ideia is that,as the file names include these words,
this function identifies which files has these words and add them as columns.}

\item{lnchar}{Integer; last nchar, default = 11.}

\item{out}{Path to the Obspack index output}

\item{verbose}{Logical to show more information}

\item{aslist}{Logical to return list of index and summary}
}
\value{
A data.frame with with an index of the obspack Globalview.

A data.frame with with an index of the obspack Globalview.
}
\description{
This function returns a data.frame index of all files
available in ObsPack.

This function returns a data.frame index of all files
available in obspack.
}
\examples{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
}
{
\dontrun{
# Do not run
obs <- system.file("data-raw", package = "rtorf")
index <- obs_summary(obs)
}
}
}
\concept{obs_summary}
