% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robspack-package.R
\docType{package}
\name{rtorf-package}
\alias{rtorf}
\alias{rtorf-package}
\title{rtorf: 'rtorf' Tools for Observations, Receptors and Footprints}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reads, process, and plots Observations such as the CH4 ObsPack GLOBALView+ data from NOAA and perform different filters and analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noaa-gml/rtorf}
  \item \url{https://noaa-gml.github.io/rtorf/}
  \item Report bugs at \url{https://github.com/noaa-gml/rtorf/issues/}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarraespinosa@colorado.edu} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

Other contributors:
\itemize{
  \item Lei Hu \email{lei.hu@noaa.gov} (\href{https://orcid.org/0000-0002-5624-3832}{ORCID}) [contributor]
}

}
\keyword{internal}
